+++ encoding/thrift/thriftrw-plugin-yarpc/gomock.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2024 Uber Technologies, Inc.
+// Copyright (c) 2025 Uber Technologies, Inc.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -33,7 +33,7 @@ const gomockTemplate = `
 <$pkgname := printf "%stest" (lower .Name)>
 package <$pkgname>
 
-<$gomock := import "github.com/golang/mock/gomock">
+<$gomock := import .MockLibrary>
 
 // MockClient implements a gomock-compatible mock client for service
 // <.Name>.