+++ x/debug/debug.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2024 Uber Technologies, Inc.
+// Copyright (c) 2025 Uber Technologies, Inc.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -137,6 +137,7 @@ var (
 			<th colspan="3">Chooser</th>
 		</tr>
 		<tr>
+			<th></th>
 			<th></th>
 			<th></th>
 			<th></th>